Don't use ssize_t

+++ src/lib/buffer.c
@@ -20,11 +20,11 @@ void fh_free_buffer(struct fh_buffer *buf)
   buf->cap = 0;
 }
 
-int fh_buf_grow(struct fh_buffer *buf, size_t add_size)
+int fh_buf_grow(struct fh_buffer *buf, int add_size)
 {
-  size_t new_size = buf->size + add_size;  // TODO: check overflow
+  int new_size = buf->size + add_size;  // TODO: check overflow
   if (new_size > buf->cap) {
-    size_t new_cap = ((new_size + 1024 + 1) / 1024) * 1024;
+    int new_cap = ((new_size + 1024 + 1) / 1024) * 1024;
     void *new_p = realloc(buf->p, new_cap);
     if (new_p == NULL)
       return -1;
@@ -36,9 +36,9 @@ int fh_buf_grow(struct fh_buffer *buf, size_t add_size)
   return 0;
 }
 
-ssize_t fh_buf_add_string(struct fh_buffer *buf, const char *str, size_t str_size)
+int fh_buf_add_string(struct fh_buffer *buf, const char *str, int str_size)
 {
-  ssize_t pos = buf->size;
+  int pos = buf->size;
   if (fh_buf_grow(buf, str_size + 1) < 0)
     return -1;
   memcpy(buf->p + pos, str, str_size);
@@ -46,9 +46,9 @@ ssize_t fh_buf_add_string(struct fh_buffer *buf, const char *str, size_t str_siz
   return pos;
 }
 
-ssize_t fh_buf_add_byte(struct fh_buffer *buf, uint8_t c)
+int fh_buf_add_byte(struct fh_buffer *buf, uint8_t c)
 {
-  ssize_t pos = buf->size;
+  int pos = buf->size;
   if (fh_buf_grow(buf, 1) < 0)
     return -1;
   buf->p[pos++] = c;
